Refactor image embedding to use float[] and pgvector
Changed image embedding representation from Double[] to float[] across Java and Python codebases for compatibility with pgvector in PostgreSQL. Added custom converters for Hibernate, updated model and service logic, and adjusted Python feature export and search to use consistent field names and float32 arrays. Also set eager fetching for related entities and improved product variant update logic.

+++ src/main/java/com/ds/project/common/entities/dto/request/ProductImageRequest.java
@@ -21,7 +21,7 @@ public class ProductImageRequest {
 
     private MultipartFile file;
 
-    private Double[] embedding;
+    private float[] embedding;
 
     private String productVariantId;
 }
